+++ attributes/apache2.rb
@@ -2,7 +2,7 @@
 # Cookbook Name:: stash
 # Attributes:: apache2
 #
-# Copyright 2012
+# Copyright 2012-2013
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,9 +22,17 @@ default['stash']['apache2']['error_log']          = ""
 default['stash']['apache2']['port']               = 80
 default['stash']['apache2']['virtual_host_alias'] = node['fqdn']
 default['stash']['apache2']['virtual_host_name']  = node['hostname']
+
 default['stash']['apache2']['ssl']['access_log']       = ""
-default['stash']['apache2']['ssl']['certificate_file'] = "/etc/pki/tls/certs/localhost.crt"
 default['stash']['apache2']['ssl']['chain_file']       = ""
 default['stash']['apache2']['ssl']['error_log']        = ""
-default['stash']['apache2']['ssl']['key_file']         = "/etc/pki/tls/private/localhost.key"
 default['stash']['apache2']['ssl']['port']             = 443
+
+case node['platform_family']
+when 'rhel'
+  default['stash']['apache2']['ssl']['certificate_file'] = "/etc/pki/tls/certs/localhost.crt"
+  default['stash']['apache2']['ssl']['key_file']         = "/etc/pki/tls/private/localhost.key"
+else
+  default['stash']['apache2']['ssl']['certificate_file'] = "/etc/ssl/certs/ssl-cert-snakeoil.pem"
+  default['stash']['apache2']['ssl']['key_file']         = "/etc/ssl/private/ssl-cert-snakeoil.key"
+end