+++ recipes/apache2.rb
@@ -2,7 +2,7 @@
 # Cookbook Name:: stash
 # Recipe:: apache2
 #
-# Copyright 2012
+# Copyright 2012-2013
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,8 +19,8 @@
 
 include_recipe "stash"
 
-node.set['apache']['listen_ports'] << node['stash']['apache2']['port'] unless node['apache']['listen_ports'].include?(node['stash']['apache2']['port'])
-node.set['apache']['listen_ports'] << node['stash']['apache2']['ssl']['port'] unless node['apache']['listen_ports'].include?(node['stash']['apache2']['ssl']['port'])
+node.set['apache']['listen_ports'] = node['apache']['listen_ports'] + [ node['stash']['apache2']['port'] ] unless node['apache']['listen_ports'].include?(node['stash']['apache2']['port'])
+node.set['apache']['listen_ports'] = node['apache']['listen_ports'] + [ node['stash']['apache2']['ssl']['port'] ] unless node['apache']['listen_ports'].include?(node['stash']['apache2']['ssl']['port'])
 
 include_recipe "apache2"
 include_recipe "apache2::mod_proxy"